@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Installer for the collective.transmogrifier package."""
 
 from setuptools import find_packages
@@ -16,26 +15,28 @@ long_description = "\n\n".join(
 
 setup(
     name="collective.transmogrifier",
-    version="2.0.1.dev0",
+    version="3.0.0.dev0",
     description="A configurable pipeline, aimed at transforming content for import and export",
     long_description=long_description,
     # Get more from https://pypi.org/classifiers/
     classifiers=[
         "Development Status :: 6 - Mature",
         "Environment :: Web Environment",
-        "Framework :: Plone",
-        "Framework :: Plone :: Addon",
-        "Framework :: Plone :: 4.3",
-        "Framework :: Plone :: 5.0",
-        "Framework :: Plone :: 5.1",
         "Framework :: Plone :: 5.2",
-        "Programming Language :: Python",
-        "Programming Language :: Python :: 2.7",
-        "Programming Language :: Python :: 3.6",
+        "Framework :: Plone :: 6.0",
+        "Framework :: Plone :: Addon",
+        "Framework :: Plone",
+        "Framework :: Zope2",
+        "Framework :: Zope3",
+        "License :: OSI Approved :: GNU General Public License (GPL)",
+        "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
+        "Operating System :: OS Independent",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
-        "Operating System :: OS Independent",
-        "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python",
+        "Topic :: Software Development :: Libraries :: Application Frameworks",
+        "Topic :: Software Development :: Libraries :: Python Modules",
     ],
     keywords="content import filtering",
     author="Jarn",
@@ -45,7 +46,6 @@ setup(
         "PyPI": "https://pypi.python.org/pypi/collective.transmogrifier",
         "Source": "https://github.com/collective/collective.transmogrifier",
         "Tracker": "https://github.com/collective/collective.transmogrifier/issues",
-        # 'Documentation': 'https://collective.transmogrifier.readthedocs.io/en/latest/',
     },
     license="GPL version 2",
     packages=find_packages("src", exclude=["ez_setup"]),
@@ -53,15 +53,17 @@ setup(
     package_dir={"": "src"},
     include_package_data=True,
     zip_safe=False,
-    python_requires="==2.7, >=3.6",
+    python_requires=">=3.7",
     install_requires=[
         "Products.CMFCore",
         "setuptools",
-        "six",
         "zope.schema",
     ],
     extras_require={
-        "test": [],
+        "test": [
+            "zest.releaser[recommended]",
+            "zope.testrunner",
+        ]
     },
     entry_points="""
     [z3c.autoinclude.plugin]
