@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
 from collective.transmogrifier.utils import Condition
@@ -7,7 +6,6 @@ from zope.interface import implementer
 from zope.interface import provider
 
 import codecs
-import six
 
 
 def _get_default_encoding(site):
@@ -18,7 +16,7 @@ def _get_default_encoding(site):
 
 @provider(ISectionBlueprint)
 @implementer(ISection)
-class CodecSection(object):
+class CodecSection:
 
     from_ = None
     from_error_handler = "strict"
@@ -72,36 +70,31 @@ class CodecSection(object):
         if from_ is None:
 
             def decode(value):
-                if six.PY2 and not isinstance(value, unicode):
-                    raise ValueError("Not a unicode string: %s" % value)
                 return value
 
         else:
             from_error_handler = self.from_error_handler
 
             def decode(value):
-                # Python 2 unicode or Python 3 bytes have the decode method
-                if six.PY2 or isinstance(value, bytes):
+                # Bytes have the decode method
+                if isinstance(value, bytes):
                     return value.decode(from_, from_error_handler)
-                # Strings in Python 3 don't have the decode method
+                # Strings don't have the decode method
                 return value.encode().decode(from_, from_error_handler)
 
         to = self.to
         if to is None:
 
             def encode(value):
-                if six.PY2 and not isinstance(value, unicode):
-                    raise ValueError("Not a unicode string: %s" % value)
                 return value
 
         else:
             to_error_handler = self.to_error_handler
 
             def encode(value):
-                # Python 3 bytes are converted to strings
-                if six.PY2 or isinstance(value, bytes):
+                # Bytes are converted to strings
+                if isinstance(value, bytes):
                     return value.encode(to, to_error_handler)
-                # After encoding in Python 3, we convert the result to string.
                 return value.encode(to, to_error_handler).decode()
 
         for item in self.previous:
