@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
 from collective.transmogrifier.utils import defaultMatcher
@@ -12,7 +11,7 @@ import csv
 
 @provider(ISectionBlueprint)
 @implementer(ISection)
-class CSVSourceSection(object):
+class CSVSourceSection:
     def __init__(self, transmogrifier, name, options, previous):
         self.previous = previous
         self.transmogrifier = transmogrifier
@@ -31,16 +30,15 @@ class CSVSourceSection(object):
 
         self.dialect = options.get("dialect", "excel")
         self.restkey = options.get("restkey", "_csvsource_rest")
-        self.fmtparam = dict(
-            (
-                key[len("fmtparam-") :],
-                Expression(value, transmogrifier, name, options)(
-                    options, key=key[len("fmtparam-") :]
-                ),
+        self.fmtparam = {
+            key[len("fmtparam-") :]: Expression(  # noQA
+                value, transmogrifier, name, options
+            )(
+                options, key=key[len("fmtparam-") :]  # noQA
             )
             for key, value in options.items()
             if key.startswith("fmtparam-")
-        )
+        }
         self.fieldnames = options.get("fieldnames")
         if self.fieldnames:
             self.fieldnames = self.fieldnames.split()
@@ -75,8 +73,7 @@ class CSVSourceSection(object):
             dialect=self.dialect,
             fieldnames=self.fieldnames,
             restkey=self.restkey,
-            **self.fmtparam
+            **self.fmtparam,
         )
-        for item in reader:
-            yield item
+        yield from reader
         file_.close()
