@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
 from collective.transmogrifier.utils import Condition
@@ -9,7 +8,7 @@ from zope.interface import provider
 
 @provider(ISectionBlueprint)
 @implementer(ISection)
-class InserterSection(object):
+class InserterSection:
     def __init__(self, transmogrifier, name, options, previous):
         self.key = Expression(options["key"], transmogrifier, name, options)
         self.value = Expression(options["value"], transmogrifier, name, options)
