@@ -1,45 +1,25 @@
-# -*- coding: utf-8 -*-
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
 from collective.transmogrifier.utils import defaultMatcher
 from collective.transmogrifier.utils import Expression
 from collective.transmogrifier.utils import resolvePackageReferenceOrFile
+from email import message_from_string
+from email.message import Message
 from zope.interface import implementer
 from zope.interface import provider
 
 import contextlib
-import io
 import logging
 import mimetypes
 import os
-import six
-import six.moves.urllib.error
-import six.moves.urllib.parse
-import six.moves.urllib.request
-
-
-# BBB: mimetools.Message doesn't exist in Python 3. Usually email.message.Message
-# is used in Python 3. In some cases it would be possible to use
-# email.message.Message in Python 2 as well. But here we call the method
-# six.moves.urllib.request.HTTPRedirectHandler.http_error_302.
-# In Python 2 the headers parameter of this method must be a mimetools.Message.
-# So, we need to use it here.
-if six.PY2:
-    from mimetools import Message
-else:
-    from email import message_from_string
-    from email.message import Message
-
-
-def get_message(initial_value=""):
-    if six.PY2:
-        return Message(io.StringIO(initial_value.decode()))
-    return message_from_string(initial_value)
+import urllib.error
+import urllib.parse
+import urllib.request
 
 
 @provider(ISectionBlueprint)
 @implementer(ISection)
-class URLOpenerSection(object):
+class URLOpenerSection:
     def __init__(self, transmogrifier, name, options, previous):
         self.previous = previous
 
@@ -69,7 +49,7 @@ class URLOpenerSection(object):
         if not os.path.isdir(self.cachedir):
             os.makedirs(self.cachedir)
         self.defaultpagename = options.get(
-            "default-page-name", ".{0}.cache".format(options["blueprint"])
+            "default-page-name", ".{}.cache".format(options["blueprint"])
         )
 
         handlers = Expression(
@@ -85,10 +65,10 @@ class URLOpenerSection(object):
         if not [
             handler
             for handler in handlers
-            if isinstance(handler, six.moves.urllib.request.HTTPRedirectHandler)
+            if isinstance(handler, urllib.request.HTTPRedirectHandler)
         ]:
             handlers.append(HTTPRedirectHandler())
-        self.opener = six.moves.urllib.request.build_opener(*handlers)
+        self.opener = urllib.request.build_opener(*handlers)
 
     def __iter__(self):
         for item in self.previous:
@@ -98,8 +78,8 @@ class URLOpenerSection(object):
                 continue
 
             url = item[key]
-            if not isinstance(url, six.moves.urllib.parse.SplitResult):
-                url = six.moves.urllib.parse.urlsplit(url)
+            if not isinstance(url, urllib.parse.SplitResult):
+                url = urllib.parse.urlsplit(url)
 
             cache = os.path.join(
                 self.cachedir,
@@ -118,7 +98,7 @@ class URLOpenerSection(object):
 
             if os.path.isfile(cache) and os.path.isfile(headers_cache):
                 self.logger.debug("Using cache: %s", cache)
-                headers = get_message(open(headers_cache))
+                headers = message_from_string(open(headers_cache))
             else:
                 if not os.path.isdir(os.path.dirname(cache)):
                     os.makedirs(os.path.dirname(cache))
@@ -148,25 +128,25 @@ class URLOpenerSection(object):
             yield item
 
 
-class HTTPDefaultErrorHandler(six.moves.urllib.request.HTTPDefaultErrorHandler):
+class HTTPDefaultErrorHandler(urllib.request.HTTPDefaultErrorHandler):
     def http_error_default(self, req, fp, code, msg, hdrs):
         if not isinstance(hdrs, Message):
-            hdrs = get_message(hdrs)
+            hdrs = message_from_string(hdrs)
         hdrs["status"] = str(code) + " " + msg
         try:
-            return six.moves.urllib.request.HTTPDefaultErrorHandler.http_error_default(
+            return urllib.request.HTTPDefaultErrorHandler.http_error_default(
                 self, req, fp, code, msg, hdrs
             )
-        except six.moves.urllib.error.URLError as error:
+        except urllib.error.URLError as error:
             if not self.section.ignore_error(self.item, error=error):
                 raise
             self.section.logger.warning("Ignoring error opening URL: %s", error)
             return error
 
 
-class HTTPRedirectHandler(six.moves.urllib.request.HTTPRedirectHandler):
+class HTTPRedirectHandler(urllib.request.HTTPRedirectHandler):
     def http_error_302(self, req, fp, code, msg, headers):
-        resp = six.moves.urllib.request.HTTPRedirectHandler.http_error_302(
+        resp = urllib.request.HTTPRedirectHandler.http_error_302(
             self, req, fp, code, msg, headers
         )
         resp.headers["redirect-status"] = headers.get(
