@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from collective.transmogrifier.transmogrifier import configuration_registry
 from zope.configuration.fields import MessageID
 from zope.configuration.fields import Path
@@ -13,29 +12,29 @@ class IRegisterConfigDirective(Interface):
     """
 
     name = DottedName(
-        title=u"Name",
-        description=u"If not specified 'default' is used.",
+        title="Name",
+        description="If not specified 'default' is used.",
         default="default",
         required=False,
     )
 
     title = MessageID(
-        title=u"Title",
-        description=u"Optional title for the pipeline configuration.",
+        title="Title",
+        description="Optional title for the pipeline configuration.",
         default=None,
         required=False,
     )
 
     description = MessageID(
-        title=u"Description",
-        description=u"Optional description for the pipeline configuration.",
+        title="Description",
+        description="Optional description for the pipeline configuration.",
         default=None,
         required=False,
     )
 
     configuration = Path(
-        title=u"Configuration",
-        description=u"The pipeline configuration file to register.",
+        title="Configuration",
+        description="The pipeline configuration file to register.",
         required=True,
     )
 
@@ -48,10 +47,10 @@ def registerConfig(
 ):
     """Add a new configuration to the registry"""
     if title is None:
-        title = u"Pipeline configuration '%s'" % name
+        title = "Pipeline configuration '%s'" % name
 
     if description is None:
-        description = u""
+        description = ""
 
     _configuration_regs.append("%s" % name)
 
